upload

+++ ForPractice/50_ChattingServer/00.Single/ChattingServer_Single/ChattingServer_Single/Main.cpp
@@ -1,6 +1,7 @@
 ﻿#include "CChattingServer.h"
 #include "CCrashDump.h"
 #include "CSystemLog.h"
+#include <locale.h>
 
 #ifndef _WINSOCKAPI_
 #define _WINSOCKAPI_
@@ -14,6 +15,8 @@ CChattingServer g_Server;
 
 int wmain(int argc, wchar_t* argv[])
 {
+    _wsetlocale(LC_ALL, L"korean");
+
     SYSLOG_DIRECTORY(L"SystemLog");
     SYSLOG_LEVEL(CSystemLog::DEBUG_LEVEL);
     LOG(L"FightGame", CSystemLog::SYSTEM_LEVEL, L"Main Thread Start\n");
@@ -22,8 +25,7 @@ int wmain(int argc, wchar_t* argv[])
     if (!g_Server.Initialize()) return 0;
     for (;;)
     {
-        if (GetAsyncKeyState(VK_SPACE))
-            g_Server.Terminate();
+        if (GetAsyncKeyState(VK_SPACE)) g_Server.Terminate();
     }
     timeEndPeriod(1);
